weather-func-app/requirements.txt
# azure-functions-worker in this file should not be included

azure-functions
azure.eventhub
azure.identity
azure.keyvault.secrets
requests
json